+++ CHANGELOG.rst
@@ -16,6 +16,9 @@ CHANGELOG
 * Parameter initializations/updates now go through layer functions initParameter and updateParameter
 * bug fix: Loading a saved model will no longer trigger parameter reinitialisations
 * bug fix: Precition accuracy used to be the same as the classification accuracy
+* Added HeWeights and ScaledVarianceWeights initializations
+* HeWeights are now the default initialization
+* bug fix: GeometricEarlyStopping's patience is now truly reset when a better score is achieved
  
 1.0.2rc:
 --------